Refactor: Extract memory and system utilities to src/analysis/utils/memory.rs
Extract memory detection and system requirements checking code from
implementation.rs (~500 lines) to a dedicated utils/memory.rs module.

Changes:
- Create src/analysis/utils/ directory structure with mod.rs, memory.rs,
  and platform.rs
- Extract memory detection functions: detect_memory_tier(), get_memory_info()
  (macOS, Linux, fallback variants), parse_vm_stat_line(), parse_vm_stat_page_size(),
  parse_meminfo_line()
- Extract memory validation: check_memory_for_parquet(),
  validate_parquet_memory_requirements(), check_system_memory_requirements()
- Extract GPU batch sizing: cap_gpu_batch_size_by_bytes(), get_work_queue_capacity()
- Extract platform setup: suppress_mesa_warnings_if_requested(),
  ensure_xdg_runtime_dir()
- Extract types: MemoryTier enum, batch size constants
- Add comprehensive tests for all extracted functions
- Update implementation.rs to import from new module
- Maintain public API: check_memory_for_parquet re-exported via analysis/mod.rs

Fixes #267
