+++ src/analysis/implementation.rs
@@ -21,6 +21,14 @@ use crate::analysis::activation::{
     logistic_activation, mish_activation, relu6_activation, softplus_activation,
     softsign_activation, tanh_activation, ActivationCandidateSpec, ACTIVATION_SPECS,
 };
+
+// Import memory and platform utilities from dedicated modules (Issue #267)
+use crate::analysis::utils::{
+    cap_gpu_batch_size_by_bytes, check_memory_for_parquet, check_system_memory_requirements,
+    detect_memory_tier, ensure_xdg_runtime_dir, get_memory_info, get_work_queue_capacity,
+    suppress_mesa_warnings_if_requested, verbose_enabled, MemoryTier, DEFAULT_GPU_BATCH_SIZE,
+    HIGH_PERF_GPU_BATCH_SIZE, LOW_MEMORY_GPU_BATCH_SIZE,
+};
 use bytemuck::{Pod, Zeroable};
 use crossbeam_channel::{bounded, Receiver, Sender};
 use once_cell::sync::OnceCell;
@@ -241,38 +249,6 @@ fn constant_source_effect_threshold_from_env() -> Option<f32> {
     }
 }
 
-/// Default GPU batch size for GPU operations.
-/// This is tuned for M1/M2/M3 which have moderate GPU core counts.
-const DEFAULT_GPU_BATCH_SIZE: usize = 512;
-
-/// Larger GPU batch size for high-performance GPUs (M4, M3 Pro/Max, etc.).
-/// M4 has significantly more GPU cores and can handle larger batches efficiently.
-const HIGH_PERF_GPU_BATCH_SIZE: usize = 1024;
-
-/// Smaller GPU batch size for memory-constrained systems.
-/// Used when available memory is low to prevent memory pressure.
-const LOW_MEMORY_GPU_BATCH_SIZE: usize = 256;
-
-/// Memory threshold (in GB) below which we use conservative settings.
-/// Systems with less than 8GB available should use low-memory mode.
-const LOW_MEMORY_THRESHOLD_GB: f64 = 8.0;
-
-/// Memory threshold (in GB) for standard settings.
-/// Systems with 8-16GB use standard settings.
-const STANDARD_MEMORY_THRESHOLD_GB: f64 = 16.0;
-
-/// Minimum total system memory (in GB) required for discovery.
-/// Systems with less than 4GB total RAM cannot reliably run GPU discovery
-/// without risking hangs from memory pressure/swap thrashing.
-const MINIMUM_TOTAL_MEMORY_GB: f64 = 4.0;
-
-/// Minimum available memory (in GB) required for discovery.
-/// If less than 1GB is available, discovery is disabled to prevent hangs.
-/// Note: macOS on Apple Silicon aggressively uses memory for caching, so
-/// "available" memory is often reported low even when plenty can be reclaimed.
-/// 1GB is sufficient since macOS can quickly reclaim cached/inactive pages.
-const MINIMUM_AVAILABLE_MEMORY_GB: f64 = 1.0;
-
 /// Minimum timeout (seconds) for the GPU work queue waiting for a response from the GPU thread.
 /// This is the OUTER timeout - if the GPU thread doesn't respond within this time,
 /// the queue gives up and returns an error.
@@ -309,25 +285,6 @@ const GPU_BUFFER_MAP_TIMEOUT_SECS: u64 =
 /// This cap is conservative by design; it trades a bit of peak throughput for stability on
 /// Apple Silicon (and makes time-bounded runs far more reliable).
 const GPU_MAX_BATCH_ALLOC_BYTES: usize = 256 * 1024 * 1024; // 256MB
-
-fn cap_gpu_batch_size_by_bytes(
-    configured_batch_size: usize,
-    max_sample_len: usize,
-    bytes_per_sample: usize,
-    max_batch_bytes: usize,
-) -> usize {
-    if configured_batch_size == 0 || max_sample_len == 0 || bytes_per_sample == 0 {
-        return configured_batch_size.max(1);
-    }
-    let bytes_per_op = max_sample_len.saturating_mul(bytes_per_sample);
-    if bytes_per_op == 0 {
-        return configured_batch_size.max(1);
-    }
-
-    // How many operations (sources) can we safely submit in one batch chunk?
-    let cap = (max_batch_bytes / bytes_per_op).max(1);
-    configured_batch_size.min(cap).max(1)
-}
 
 /// Timeout for GPU thread initialisation (in seconds).
 /// GPU device creation should be fast; if it takes longer, something is wrong.
@@ -562,281 +519,7 @@ fn detect_gpu_tier(adapter_info: &wgpu::AdapterInfo) -> GpuPerformanceTier {
     GpuPerformanceTier::Unknown
 }
 
-/// Memory tier for adaptive configuration.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum MemoryTier {
-    /// Less than 8GB available - use conservative settings
-    Low,
-    /// 8-16GB available - use standard settings
-    Standard,
-    /// More than 16GB available - use aggressive settings
-    High,
-}
-
-/// Detect available system memory and categorise into tiers.
-/// This is cached for the lifetime of the process.
-fn detect_memory_tier() -> MemoryTier {
-    use std::sync::OnceLock;
-    static TIER: OnceLock<MemoryTier> = OnceLock::new();
-
-    *TIER.get_or_init(|| {
-        let (available, total) = get_memory_info();
-        let available_gb = available as f64 / (1024.0 * 1024.0 * 1024.0);
-        let total_gb = total as f64 / (1024.0 * 1024.0 * 1024.0);
-
-        let memory_tier = if available_gb < LOW_MEMORY_THRESHOLD_GB {
-            MemoryTier::Low
-        } else if available_gb < STANDARD_MEMORY_THRESHOLD_GB {
-            MemoryTier::Standard
-        } else {
-            MemoryTier::High
-        };
-
-        // Log memory info once
-        let tier_str = match memory_tier {
-            MemoryTier::Low => "low",
-            MemoryTier::Standard => "standard",
-            MemoryTier::High => "high",
-        };
-        eprintln!(
-            "[NEAT-AI-Discovery] Memory: {available_gb:.1}GB available / {total_gb:.1}GB total | Tier: {tier_str}"
-        );
-
-        memory_tier
-    })
-}
-
-/// Get memory information from the OS.
-/// Returns (available_bytes, total_bytes).
-#[cfg(target_os = "macos")]
-fn get_memory_info() -> (u64, u64) {
-    use std::process::Command;
-
-    // Get total memory from sysctl
-    let total = Command::new("sysctl")
-        .args(["-n", "hw.memsize"])
-        .output()
-        .ok()
-        .and_then(|o| {
-            String::from_utf8_lossy(&o.stdout)
-                .trim()
-                .parse::<u64>()
-                .ok()
-        })
-        .unwrap_or(8 * 1024 * 1024 * 1024); // Default 8GB
-
-    // Get page size and free/inactive pages from vm_stat
-    let vm_stat = Command::new("vm_stat").output().ok();
-    let available = vm_stat
-        .map(|o| {
-            let output = String::from_utf8_lossy(&o.stdout);
-            // Parse page size from vm_stat header - handles both Apple Silicon (16KB)
-            // and Intel Macs (4KB) correctly
-            let page_size = parse_vm_stat_page_size(&output);
-
-            let mut free_pages: u64 = 0;
-            let mut inactive_pages: u64 = 0;
-            let mut purgeable_pages: u64 = 0;
-
-            for line in output.lines() {
-                if line.starts_with("Pages free:") {
-                    free_pages = parse_vm_stat_line(line);
-                } else if line.starts_with("Pages inactive:") {
-                    inactive_pages = parse_vm_stat_line(line);
-                } else if line.starts_with("Pages purgeable:") {
-                    purgeable_pages = parse_vm_stat_line(line);
-                }
-            }
-
-            // Available = free + inactive + purgeable (memory that can be reclaimed)
-            (free_pages + inactive_pages + purgeable_pages) * page_size
-        })
-        .unwrap_or(total / 2); // Default to half of total
-
-    (available, total)
-}
-
-#[cfg(target_os = "macos")]
-fn parse_vm_stat_line(line: &str) -> u64 {
-    line.split(':')
-        .nth(1)
-        .and_then(|s| s.trim().trim_end_matches('.').parse::<u64>().ok())
-        .unwrap_or(0)
-}
-
-/// Parse the page size from vm_stat output's header line.
-/// Example: "Mach Virtual Memory Statistics: (page size of 16384 bytes)"
-/// Returns the page size in bytes, or a default based on architecture.
-///
-/// Apple Silicon uses 16KB pages, Intel Macs use 4KB pages.
-/// Parsing dynamically ensures correct memory calculations on both.
-#[cfg(target_os = "macos")]
-fn parse_vm_stat_page_size(output: &str) -> u64 {
-    // Default page sizes by architecture
-    // Apple Silicon (ARM64): 16384 bytes (16KB)
-    // Intel (x86_64): 4096 bytes (4KB)
-    #[cfg(target_arch = "aarch64")]
-    let default_page_size: u64 = 16384;
-    #[cfg(not(target_arch = "aarch64"))]
-    let default_page_size: u64 = 4096;
-
-    // Parse from first line: "Mach Virtual Memory Statistics: (page size of XXXX bytes)"
-    output
-        .lines()
-        .next()
-        .and_then(|first_line| {
-            // Find "page size of " and extract the number before " bytes"
-            let marker = "page size of ";
-            first_line.find(marker).and_then(|start| {
-                let after_marker = &first_line[start + marker.len()..];
-                after_marker
-                    .split_whitespace()
-                    .next()
-                    .and_then(|num_str| num_str.parse::<u64>().ok())
-            })
-        })
-        .unwrap_or(default_page_size)
-}
-
-/// Get memory information from the OS (Linux version).
-/// Works on Ubuntu, AWS Linux (Amazon Linux 2/2023), and other Linux distributions.
-/// All Linux systems expose memory info via /proc/meminfo.
-#[cfg(target_os = "linux")]
-fn get_memory_info() -> (u64, u64) {
-    use std::fs;
-
-    let meminfo = fs::read_to_string("/proc/meminfo").unwrap_or_default();
-    let mut total: u64 = 8 * 1024 * 1024; // Default 8GB in KB
-    let mut available: u64 = 4 * 1024 * 1024; // Default 4GB in KB
-
-    for line in meminfo.lines() {
-        if line.starts_with("MemTotal:") {
-            // Only update if parsing succeeds - don't overwrite defaults with 0
-            if let Some(value) = parse_meminfo_line(line) {
-                total = value;
-            }
-        } else if line.starts_with("MemAvailable:") {
-            // Only update if parsing succeeds - don't overwrite defaults with 0
-            if let Some(value) = parse_meminfo_line(line) {
-                available = value;
-            }
-        }
-    }
-
-    // Convert KB to bytes
-    (available * 1024, total * 1024)
-}
-
-/// Parse a memory value from a /proc/meminfo line.
-/// Returns None if the line is malformed (missing or non-numeric value).
-/// Example: "MemTotal:       16384000 kB" -> Some(16384000)
-#[cfg(target_os = "linux")]
-fn parse_meminfo_line(line: &str) -> Option<u64> {
-    line.split_whitespace()
-        .nth(1)
-        .and_then(|s| s.parse::<u64>().ok())
-}
-
-/// Fallback for other platforms (Windows, FreeBSD, etc.).
-/// Returns conservative defaults since we don't have platform-specific memory detection.
-/// GPU discovery will still work via wgpu (DirectX 12 on Windows, Vulkan elsewhere).
-#[cfg(not(any(target_os = "macos", target_os = "linux")))]
-fn get_memory_info() -> (u64, u64) {
-    // Conservative defaults: 8GB total, 4GB available
-    // These are safe values that won't trigger low-memory protections on modern machines
-    (4 * 1024 * 1024 * 1024, 8 * 1024 * 1024 * 1024)
-}
-
-/// Check if there's enough memory to load a parquet file.
-///
-/// Parquet files are compressed; in-memory representation is typically 2-4x larger.
-/// This function estimates memory needed and returns an error if insufficient.
-///
-/// We use conservative checks to avoid memory pressure that causes the system to
-/// become unresponsive (stuck with low CPU/GPU, eventually OOM killed):
-/// 1. Require 3× file size for in-memory representation
-/// 2. Require at least 1GB headroom after loading (for GPU buffers, etc.)
-/// 3. Don't use more than 50% of total RAM for parquet data
-///
-/// This is a public function so it can be used by focus.rs and analysis.rs.
-pub fn check_memory_for_parquet(parquet_file: &str) -> Result<()> {
-    const MEMORY_MULTIPLIER: f64 = 3.0;
-    const BYTES_PER_GB: f64 = 1024.0 * 1024.0 * 1024.0;
-    const MIN_HEADROOM_GB: f64 = 1.0; // Keep at least 1GB free for GPU/system
-    const MAX_MEMORY_FRACTION: f64 = 0.5; // Don't use more than 50% of total RAM
-
-    let file_size_bytes = std::fs::metadata(parquet_file)
-        .map(|m| m.len())
-        .unwrap_or(0);
-    let file_size_gb = file_size_bytes as f64 / BYTES_PER_GB;
-    let estimated_memory_gb = file_size_gb * MEMORY_MULTIPLIER;
-
-    let (available_bytes, total_bytes) = get_memory_info();
-    let available_gb = available_bytes as f64 / BYTES_PER_GB;
-    let total_gb = total_bytes as f64 / BYTES_PER_GB;
-
-    let file_size_mb = file_size_bytes as f64 / (1024.0 * 1024.0);
-    let estimated_mb = estimated_memory_gb * 1024.0;
-    let available_mb = available_gb * 1024.0;
-
-    // Check 1: Do we have enough available memory (with headroom)?
-    let memory_needed_with_headroom = estimated_memory_gb + MIN_HEADROOM_GB;
-    if memory_needed_with_headroom > available_gb {
-        return Err(anyhow::anyhow!(
-            "Insufficient memory to load parquet file.\n\
-             • Parquet file: {file_size_mb:.0} MB ({parquet_file})\n\
-             • Estimated memory needed: {estimated_mb:.0} MB (parquet decompresses to ~3x in memory)\n\
-             • Plus 1 GB headroom for GPU buffers and system\n\
-             • Available memory: {available_mb:.0} MB / {total_gb:.1} GB total\n\n\
-             Suggestions:\n\
-             • Close other applications to free memory\n\
-             • Reduce sample rate (e.g., discoverySampleRate: 0.02) to create smaller parquet files\n\
-             • Reduce recording time (discoveryRecordTimeOutMinutes)\n\
-             • Use a machine with more RAM (16GB+ recommended for large creatures)"
-        ));
-    }
-
-    // Check 2: Don't exceed 50% of total RAM (prevents system instability)
-    let max_allowed_gb = total_gb * MAX_MEMORY_FRACTION;
-    if estimated_memory_gb > max_allowed_gb {
-        let max_parquet_mb = (max_allowed_gb / MEMORY_MULTIPLIER) * 1024.0;
-        return Err(anyhow::anyhow!(
-            "Parquet file too large for system memory.\n\
-             • Parquet file: {file_size_mb:.0} MB ({parquet_file})\n\
-             • Estimated memory needed: {estimated_mb:.0} MB ({:.0}% of {total_gb:.0} GB total RAM)\n\
-             • Maximum safe usage: {:.0} MB (50% of RAM = {max_allowed_gb:.1} GB)\n\
-             • Maximum parquet file size for this machine: {max_parquet_mb:.0} MB\n\n\
-             Suggestions:\n\
-             • Reduce sample rate (e.g., discoverySampleRate: 0.02)\n\
-             • Reduce recording time (discoveryRecordTimeOutMinutes)\n\
-             • Use a machine with more RAM (16GB+ recommended for large creatures)",
-            (estimated_memory_gb / total_gb) * 100.0,
-            max_allowed_gb * 1024.0
-        ));
-    }
-
-    // Log memory usage for large files (helps diagnose issues)
-    if file_size_gb > 0.5 {
-        let usage_percent = (estimated_memory_gb / total_gb) * 100.0;
-        eprintln!(
-            "[NEAT-AI-Discovery] Loading {file_size_mb:.0} MB parquet file \
-             (estimated {estimated_mb:.0} MB in memory = {usage_percent:.0}% of RAM, \
-             {available_mb:.0} MB available)"
-        );
-    }
-
-    Ok(())
-}
-
-/// Get the GPU work queue capacity based on system resources.
-/// Lower capacity = more backpressure = less memory usage.
-fn get_work_queue_capacity() -> usize {
-    match detect_memory_tier() {
-        MemoryTier::Low => 4,      // Aggressive backpressure
-        MemoryTier::Standard => 8, // Moderate backpressure
-        MemoryTier::High => 16,    // Allow more parallelism
-    }
-}
+// Memory detection functions moved to crate::analysis::utils::memory (Issue #267)
 
 /// Check if the system meets minimum requirements for GPU discovery.
 ///
@@ -848,37 +531,17 @@ fn get_work_queue_capacity() -> usize {
 /// Returns `None` if requirements ARE met (continue with GPU check).
 fn check_minimum_system_requirements() -> Option<GpuAvailabilityResult> {
     let (available, total) = get_memory_info();
-    let available_gb = available as f64 / (1024.0 * 1024.0 * 1024.0);
-    let total_gb = total as f64 / (1024.0 * 1024.0 * 1024.0);
 
-    // Check total system memory
-    if total_gb < MINIMUM_TOTAL_MEMORY_GB {
+    // Use the utility function from the memory module
+    if let Some(reason) = check_system_memory_requirements(available, total) {
+        let available_gb = available as f64 / (1024.0 * 1024.0 * 1024.0);
+        let total_gb = total as f64 / (1024.0 * 1024.0 * 1024.0);
         eprintln!(
-            "[NEAT-AI-Discovery] System has only {total_gb:.1}GB total RAM (minimum: {MINIMUM_TOTAL_MEMORY_GB}GB). Discovery disabled."
+            "[NEAT-AI-Discovery] Memory check failed: {available_gb:.1}GB available / {total_gb:.1}GB total. Discovery disabled."
         );
         return Some(GpuAvailabilityResult {
             available: false,
-            reason: Some(format!(
-                "Insufficient system memory: {total_gb:.1}GB total (minimum: {MINIMUM_TOTAL_MEMORY_GB}GB required). \
-                 Discovery is disabled to prevent hangs on memory-constrained machines. \
-                 Evolution will continue without discovery."
-            )),
-            is_error: false, // Not an error - graceful disable
-        });
-    }
-
-    // Check available memory
-    if available_gb < MINIMUM_AVAILABLE_MEMORY_GB {
-        eprintln!(
-            "[NEAT-AI-Discovery] Only {available_gb:.1}GB memory available (minimum: {MINIMUM_AVAILABLE_MEMORY_GB}GB). Discovery disabled."
-        );
-        return Some(GpuAvailabilityResult {
-            available: false,
-            reason: Some(format!(
-                "Insufficient available memory: {available_gb:.1}GB available (minimum: {MINIMUM_AVAILABLE_MEMORY_GB}GB required). \
-                 Discovery is disabled to prevent hangs from memory pressure. \
-                 Try closing other applications or reboot to free memory."
-            )),
+            reason: Some(reason),
             is_error: false, // Not an error - graceful disable
         });
     }
@@ -1493,95 +1156,8 @@ mod deadline_override {
     }
 }
 
-/// Check if verbose logging is enabled. Result is cached for performance.
-/// Set `NEAT_AI_DISCOVERY_VERBOSE=1` to enable verbose logging.
-/// Check if verbose logging is enabled via NEAT_AI_DISCOVERY_VERBOSE environment variable.
-/// This is public so it can be used by focus.rs and other modules.
-pub fn verbose_enabled() -> bool {
-    use std::sync::OnceLock;
-    static VERBOSE: OnceLock<bool> = OnceLock::new();
-    *VERBOSE.get_or_init(|| std::env::var("NEAT_AI_DISCOVERY_VERBOSE").is_ok())
-}
-
-/// Suppress Mesa/libEGL debug warnings on Linux.
-///
-/// When wgpu initialises on Linux, it probes multiple GPU backends (EGL, Vulkan, etc.).
-/// If the user lacks permission to access `/dev/dri/renderD*` or `/dev/dri/card*` devices,
-/// libEGL emits warnings like "failed to open /dev/dri/renderD128: Permission denied".
-///
-/// These warnings are often benign if wgpu finds an alternative backend (e.g., Vulkan via
-/// a different ICD loader). This function suppresses the warnings by setting environment
-/// variables that quiet Mesa's debug output.
-///
-/// Set `NEAT_AI_DISCOVERY_QUIET_GPU=1` to enable suppression.
-#[cfg(target_os = "linux")]
-fn suppress_mesa_warnings_if_requested() {
-    use std::env;
-    use std::sync::Once;
-
-    static INIT: Once = Once::new();
-    INIT.call_once(|| {
-        if env::var("NEAT_AI_DISCOVERY_QUIET_GPU").is_ok() {
-            // Suppress EGL debug messages (these cause "failed to open /dev/dri/..." warnings)
-            if env::var("EGL_LOG_LEVEL").is_err() {
-                // SAFETY: single-threaded at this point (Once guard) and before GPU init
-                unsafe { env::set_var("EGL_LOG_LEVEL", "fatal") };
-            }
-
-            // Suppress Mesa GLSL shader cache warnings
-            if env::var("MESA_GLSL_CACHE_DISABLE").is_err() {
-                unsafe { env::set_var("MESA_GLSL_CACHE_DISABLE", "true") };
-            }
-
-            // Suppress general Mesa debug output
-            if env::var("MESA_DEBUG").is_err() {
-                unsafe { env::set_var("MESA_DEBUG", "silent") };
-            }
-        }
-    });
-}
-
-#[cfg(not(target_os = "linux"))]
-fn suppress_mesa_warnings_if_requested() {
-    // No-op on non-Linux platforms
-}
-
-/// Ensure XDG_RUNTIME_DIR is set on Linux (required by wgpu on Wayland).
-///
-/// This function uses `Once` for thread-safe one-time initialisation. It's safe to
-/// call from multiple threads concurrently - only the first call will set the
-/// environment variable, and subsequent calls are no-ops.
-///
-/// Must be called before any GPU initialisation (wgpu Instance creation).
-#[cfg(target_os = "linux")]
-fn ensure_xdg_runtime_dir() {
-    use std::env;
-    use std::sync::Once;
-
-    static INIT: Once = Once::new();
-    INIT.call_once(|| {
-        if env::var("XDG_RUNTIME_DIR").is_err() {
-            // Create a temporary runtime directory if XDG_RUNTIME_DIR is not set
-            if let Ok(temp_dir) = std::env::temp_dir().canonicalize() {
-                let runtime_dir = temp_dir.join("neat-ai-discovery-runtime");
-                if let Err(e) = std::fs::create_dir_all(&runtime_dir) {
-                    eprintln!("[NEAT-AI-Discovery] Warning: Failed to create XDG_RUNTIME_DIR at {runtime_dir:?}: {e}");
-                } else {
-                    // SAFETY: Inside Once::call_once, so guaranteed single-threaded execution.
-                    // Called before any GPU init.
-                    unsafe {
-                        env::set_var("XDG_RUNTIME_DIR", runtime_dir.to_string_lossy().as_ref());
-                    }
-                }
-            }
-        }
-    });
-}
-
-#[cfg(not(target_os = "linux"))]
-fn ensure_xdg_runtime_dir() {
-    // No-op on non-Linux platforms
-}
+// verbose_enabled, suppress_mesa_warnings_if_requested, and ensure_xdg_runtime_dir
+// moved to crate::analysis::utils (Issue #267)
 
 /// Safely create a wgpu Instance, avoiding panics from backend probing.
 ///
