+++ src/analysis/utils/mod.rs
@@ -2,6 +2,25 @@
 //!
 //! This module contains helper functions for memory checks, deadline handling,
 //! activation functions, and other utilities used across analysis modules.
+//!
+//! ## Module Structure
+//!
+//! - `memory` - Memory detection and system requirements checking (Issue #267)
+//! - `platform` - Platform-specific setup (Linux XDG, Mesa warnings) (Issue #267)
+
+pub mod memory;
+pub mod platform;
+
+// Re-export key memory functions for convenience
+pub use memory::{
+    cap_gpu_batch_size_by_bytes, categorise_memory_tier, check_memory_for_parquet,
+    check_system_memory_requirements, detect_memory_tier, get_memory_info, get_work_queue_capacity,
+    get_work_queue_capacity_for_tier, validate_parquet_memory_requirements, MemoryTier,
+    DEFAULT_GPU_BATCH_SIZE, HIGH_PERF_GPU_BATCH_SIZE, LOW_MEMORY_GPU_BATCH_SIZE,
+};
+
+// Re-export platform setup functions
+pub use platform::{ensure_xdg_runtime_dir, suppress_mesa_warnings_if_requested};
 
 /// Check if verbose logging is enabled. Result is cached for performance.
 /// Set `NEAT_AI_DISCOVERY_VERBOSE=1` to enable verbose logging.
@@ -62,7 +81,7 @@ pub(crate) fn sensible_bias_abs_max_for_squash(_squash: &str) -> f32 {
 /// it's simply not worth returning because TypeScript will cache the failure and
 /// never re-try it.
 ///
-/// Note: This is applied AFTER the "Extreme → Conservative/Gentle Nudge" pairing, so
+/// Note: This is applied AFTER the "Extreme -> Conservative/Gentle Nudge" pairing, so
 /// unsafe originals can be dropped while still keeping safe variants.
 #[doc(hidden)]
 pub fn filter_candidates_to_sensible_ranges(
@@ -297,17 +316,3 @@ pub fn pair_extreme_candidates_with_conservative_variants(
 
     output
 }
-
-// TODO: Move other utility functions from impl.rs here:
-// - check_memory_for_parquet
-// - get_memory_info (platform-specific)
-// - parse_vm_stat_line, parse_vm_stat_page_size (macOS)
-// - parse_meminfo_line (Linux)
-// - detect_memory_tier
-// - build_deadline, deadline_passed, calculate_effective_timeout_ms
-// - log_analysis_start, log_analysis_timeout
-// - wait_for_buffer_map, wait_for_buffer_maps_batch
-// - calculate_gpu_batch_timeout
-// - All activation functions (identity_activation, tanh_activation, etc.)
-// - is_threshold_activation, has_sufficient_output_variance
-// - activation_name_to_gpu_id, ACTIVATION_SPECS
